Moved create channel tests
